solved xml parsing issues, added constants and settings loading, fixed Configuration classes

+++ GameMaster/GameMaster.cs
@@ -80,7 +80,7 @@ namespace GameMaster
 
         private void InitBoard(GameMasterSettingsGameDefinition settings)
         {
-            actualBoard = new Board(settings.BoardWidth, settings.TaskAreaLength, settings.GoalAreaLength);
+            actualBoard = new Board((uint)settings.BoardWidth, settings.TaskAreaLength, settings.GoalAreaLength);
             PlaceInitialPieces(settings.InitialNumberOfPieces);
             PlaceInitialGoals(settings.Goals);
         }
